bugFixes 1.0

+++ service/reservService.ts
@@ -39,21 +39,22 @@ export class ReservService {
     try {
       if (this.collection && this.collection2) {
         const limite = await this.collection2.find({isn_id: reserva.isn_id_cop}).project({_id: 0}).toArray();
-        if (!limite[0].reservado || !limite[0].emprestado) {
+        if(!limite[0].emprestado && !limite[0].reservado) {
           reserva.dt_ret = new Date();
           reserva.dt_devol = new Date();
           reserva.dt_reserva = new Date();
 
           const r = reserva.dt_reserva.getDate() + 3;
           reserva.dt_ret.setDate(r);
-          const s = reserva.dt_ret.getDate() + limite;
+          const s = reserva.dt_ret.getDate() + limite[0].categoria.dias_limite;
           reserva.dt_devol.setDate(s);
           const reservado = true;
           await this.collection2.updateOne({isn_id: reserva.isn_id_cop}, {$set: {reservado}});
           const {acknowledged} = await this.collection.insertOne(reserva);
           return acknowledged;
+        } else {
+          return false;
         }
-        return false;
       }
     } catch (error) {
       throw error;
@@ -62,7 +63,7 @@ export class ReservService {
 
   async deleteReserv(id: string) {
     try {
-      if (this.collection) {
+      if (this.collection && this.collection2) {
         const reservado = false;
         await this.collection2.updateOne({isn_id: id}, {$set: {reservado}});
         const {acknowledged} = await this.collection.deleteOne({isn_id_cop: id});
